requirements.txt
pandas
numpy
requests
PyYAML
websockets
watchdog
python-dotenv
alpaca-py
redis

